requirements.txt
agency-swarm==0.4.4
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
python-dotenv==1.0.1 